+++ MatchDataManager.Api/Controllers/TeamsController.cs
@@ -1,6 +1,8 @@
+using MatchDataManager.Api.MatchData;
 using MatchDataManager.Api.Models;
 using MatchDataManager.Api.Repositories;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 namespace MatchDataManager.Api.Controllers;
 
@@ -8,42 +10,75 @@ namespace MatchDataManager.Api.Controllers;
 [Route("[controller]")]
 public class TeamsController : ControllerBase
 {
+    private readonly MatchDataDbContext _context;
+
+    public TeamsController(MatchDataDbContext context)
+    {
+        _context = context;
+    }
     [HttpPost]
-    public IActionResult AddTeam(Team team)
+    public async Task<ActionResult<Team>> AddTeam(Team team)
     {
         TeamsRepository.AddTeam(team);
+        _context.Team.Add(team);
+        await _context.SaveChangesAsync();
         return CreatedAtAction(nameof(GetById), new {id = team.Id}, team);
     }
 
     [HttpDelete]
-    public IActionResult DeleteTeam(Guid teamId)
+    public async Task<IActionResult> DeleteTeam(Guid teamId)
     {
+        var deletedTeam = await _context.Team.FindAsync(teamId);
+        
+        if (deletedTeam == null)
+        {
+            return NotFound();
+        }
         TeamsRepository.DeleteTeam(teamId);
+        _context.Team.Remove(deletedTeam);
         return NoContent();
     }
 
     [HttpGet]
-    public IActionResult Get()
+    public async Task<ActionResult<IEnumerable<Team>>> Get()
     {
-        return Ok(TeamsRepository.GetAllTeams());
+        return Ok(await _context.Team.ToListAsync());
     }
 
     [HttpGet("{id:guid}")]
-    public IActionResult GetById(Guid id)
+    public async Task<ActionResult<Team>> GetById(Guid id)
     {
-        var location = TeamsRepository.GetTeamById(id);
-        if (location is null)
+        var teamInDataBase = await _context.Team.FindAsync(id);
+        //var location = TeamsRepository.GetTeamById(id);
+        if (teamInDataBase is null)
         {
             return NotFound();
         }
 
-        return Ok(location);
+        return Ok(teamInDataBase);
     }
 
     [HttpPut]
-    public IActionResult UpdateTeam(Team team)
+    public async Task<IActionResult> UpdateTeam(Team team)
     {
         TeamsRepository.UpdateTeam(team);
-        return Ok(team);
+        _context.Entry(team).State = EntityState.Modified;
+
+        try
+        {
+            await _context.SaveChangesAsync();
+        }
+        catch(DbUpdateConcurrencyException)
+        {
+            if (TeamsRepository.GetTeamById(team.Id) != null)
+            {
+                return NotFound();
+            }
+            else
+            {
+                throw;
+            }
+        }
+        return NoContent();
     }
 }
